pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "autombist"
version = "0.3.0"
description = "MBIST wrapper and fault-flow generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Jinja2>=3.1",
  "PyYAML>=6.0",
  "typer[all]>=0.12",
]

[project.scripts]
autombist = "autombist.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/autombist"]
